requirements.txt
python-dotenv==1.0.0
google-generativeai==0.3.2
polib==1.2.0
tqdm==4.66.1
python-dateutil==2.8.2
